+++ lib/wrapture/function_spec.rb
@@ -228,7 +228,9 @@ module Wrapture
         yield
         yield "  #{wrapped_call_expression};"
         yield
-        @wrapped.error_check { |line| yield "  #{line}" }
+        @wrapped.error_check(return_val: return_variable) do |line|
+          yield "  #{line}"
+        end
       else
         yield "  #{wrapped_call_expression};"
       end
@@ -312,7 +314,7 @@ module Wrapture
     def locals
       yield 'va_list variadic_args;' if variadic?
 
-      if @wrapped.error_check?
+      if @wrapped.use_return? && !@constructor
         wrapped_type = resolve_type(@wrapped.return_val_type)
         yield "#{wrapped_type.variable('return_val')};"
       end
@@ -326,6 +328,15 @@ module Wrapture
         "new#{@spec['return']['type'].chomp('*').strip} ( #{value} )"
       else
         resolved_return.cast_expression(value)
+      end
+    end
+
+    # The name of the variable holding the return value.
+    def return_variable
+      if @constructor
+        'this->equivalent'
+      else
+        'return_val'
       end
     end